+++ ByCalc/Operator.cs
@@ -9,33 +9,33 @@ public interface IOperator
 {
     string Symbol { get; }
     int Precedence { get; }     /// highter = evaluated first
-    double Calculate(double left, double right);
+    decimal Calculate(decimal left, decimal right);
 }
 
 public class AdditionOperator : IOperator
 {
     public string Symbol => "+";
     public int Precedence => 1;
-    public double Calculate(double left, double right) => left + right;
+    public decimal Calculate(decimal left, decimal right) => left + right;
 }
 
 public class SubtractionOperator : IOperator
 {
     public string Symbol => "-";
     public int Precedence => 1;
-    public double Calculate(double left, double right) => left - right;
+    public decimal Calculate(decimal left, decimal right) => left - right;
 }
 
 public class MultiplicationOperator : IOperator
 {
     public string Symbol => "*";
     public int Precedence => 2;
-    public double Calculate(double left, double right) => left * right;
+    public decimal Calculate(decimal left, decimal right) => left * right;
 }
 
 public class DivisionOperator : IOperator
 {
     public string Symbol => "/";
     public int Precedence => 2;
-    public double Calculate(double left, double right) => left / right;
+    public decimal Calculate(decimal left, decimal right) => left / right;
 }
